refactor(shell): move Rendered/Skipped into shell/render.rs
